requirements.txt
streamlit
pandas
scikit-learn
numpy
scipy
passlib[bcrypt]
requests
requests-cache
joblib
python-dotenv
